Move compression image from Utils
